+++ packages/thumbprint-react/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+-   [Major] Rename `Avatar` component to `UserAvatar`. (#25)
 -   [Patch] Redesign the avatar with initials design to use dynamic colours. (#25)
 
 ## 2.1.0 - 2019-06-04